2.1 改了一个小bug

+++ back-end/XHU_HeadLine/src/main/java/com/xhu/headline_server/entity/User.java
@@ -23,9 +23,10 @@ public class User {
     private String updateTime;
     private String delete;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
+    // dzh
 
     public void setId(Long id) {
         this.id = id;
